feat(#23): add CompileMojo for consistency
