+++ src/main/java/org/eolang/opeo/DecompileMojo.java
@@ -23,20 +23,23 @@
  */
 package org.eolang.opeo;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+import com.jcabi.log.Logger;
+import org.apache.maven.plugin.AbstractMojo;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
 
 /**
- * Unit test for {@link OpeoMojo}.
+ * Decompiles bytecode in EO representation into high-level EO representation.
+ * The input for this mojo produced by the "jeo-maven-plugin":
+ * <a href="https://github.com/objectionary/jeo-maven-plugin">link</a>
+ *
  * @since 0.1
  */
-class OpeoMojoTest {
+@Mojo(name = "decompile", defaultPhase = LifecyclePhase.PROCESS_CLASSES)
+public final class DecompileMojo extends AbstractMojo {
 
-    @Test
-    void createsMojoWithoutProblems() {
-        Assertions.assertDoesNotThrow(
-            () -> new OpeoMojo().execute(),
-            String.format("Can't create %s instance and execute it", OpeoMojo.class)
-        );
+    @Override
+    public void execute() {
+        Logger.info(this, "opeo-maven-plugin: started decompiling bytecode into EO");
     }
 }
